Move all imports handling into Query
